+++ extension_config.cmake
@@ -4,7 +4,6 @@
 duckdb_extension_load(boilstream
     SOURCE_DIR ${CMAKE_CURRENT_LIST_DIR}
     LOAD_TESTS
-    REQUIRES httpfs quack
     LINKED_LIBS "${CMAKE_CURRENT_LIST_DIR}/opaque-client/target/wasm32-unknown-emscripten/release/libopaque_client.a"
 )